refactor: create abstract UserDataRepository containing all user-data related methods and add stubs for libretube sync repository
