+++ app/src/main/java/com/github/libretube/api/SubscriptionHelper.kt
@@ -2,51 +2,36 @@ package com.github.libretube.api
 
 import com.github.libretube.api.obj.Subscription
 import com.github.libretube.db.obj.SubscriptionsFeedItem
-import com.github.libretube.helpers.PreferenceHelper
-import com.github.libretube.repo.AccountSubscriptionsRepository
 import com.github.libretube.repo.FeedProgress
 import com.github.libretube.repo.FeedRepository
-import com.github.libretube.repo.LocalFeedRepository
-import com.github.libretube.repo.PipedAccountFeedRepository
-import com.github.libretube.repo.PipedLocalSubscriptionsRepository
-import com.github.libretube.repo.SubscriptionsRepository
+import com.github.libretube.repo.UserDataRepository
+import com.github.libretube.repo.UserDataRepositoryHelper
 
 object SubscriptionHelper {
-    /**
-     * The maximum number of channel IDs that can be passed via a GET request for fetching
-     * the subscriptions list and the feed
-     */
-    const val GET_SUBSCRIPTIONS_LIMIT = 100
-
-    private val token get() = PreferenceHelper.getToken()
-    private val subscriptionsRepository: SubscriptionsRepository
-        get() = when {
-            token.isNotEmpty() -> AccountSubscriptionsRepository()
-            else -> PipedLocalSubscriptionsRepository()
-        }
+    @Suppress("DEPRECATION")
+    private val userDataRepository: UserDataRepository
+        get() = UserDataRepositoryHelper.userDataRepository
+    @Suppress("DEPRECATION")
     private val feedRepository: FeedRepository
-        get() = when {
-            token.isNotEmpty() -> PipedAccountFeedRepository()
-            else -> LocalFeedRepository()
-        }
+        get() = UserDataRepositoryHelper.feedRepository
 
     suspend fun subscribe(
         channelId: String, name: String, uploaderAvatar: String?, verified: Boolean
-    ) = subscriptionsRepository.subscribe(channelId, name, uploaderAvatar, verified)
+    ) = userDataRepository.subscribe(channelId, name, uploaderAvatar, verified)
 
     suspend fun unsubscribe(channelId: String) {
-        subscriptionsRepository.unsubscribe(channelId)
+        userDataRepository.unsubscribe(channelId)
         // remove videos from (local) feed
         feedRepository.removeChannel(channelId)
     }
-    suspend fun isSubscribed(channelId: String) = subscriptionsRepository.isSubscribed(channelId)
+    suspend fun isSubscribed(channelId: String) = userDataRepository.isSubscribed(channelId)
     suspend fun importSubscriptions(newChannels: List<String>) =
-        subscriptionsRepository.importSubscriptions(newChannels)
+        userDataRepository.importSubscriptions(newChannels)
 
     suspend fun getSubscriptions() =
-        subscriptionsRepository.getSubscriptions().sortedBy { it.name.lowercase() }
+        userDataRepository.getSubscriptions().sortedBy { it.name.lowercase() }
 
-    suspend fun getSubscriptionChannelIds() = subscriptionsRepository.getSubscriptionChannelIds()
+    suspend fun getSubscriptionChannelIds() = userDataRepository.getSubscriptionChannelIds()
     suspend fun getFeed(forceRefresh: Boolean, onProgressUpdate: (FeedProgress) -> Unit = {}) =
         feedRepository.getFeed(forceRefresh, onProgressUpdate)
 
@@ -54,5 +39,5 @@ object SubscriptionHelper {
         feedRepository.submitFeedItemChange(feedItem)
 
     suspend fun submitSubscriptionChannelInfosChanged(subscriptions: List<Subscription>) =
-        subscriptionsRepository.submitSubscriptionChannelInfosChanged(subscriptions)
+        userDataRepository.submitSubscriptionChannelInfosChanged(subscriptions)
 }
