+++ app/src/main/java/com/github/libretube/repo/LocalUserDataRepository.kt
@@ -1,0 +1,256 @@
+package com.github.libretube.repo
+
+import android.util.Log
+import com.github.libretube.api.MediaServiceRepository
+import com.github.libretube.api.PlaylistsHelper
+import com.github.libretube.api.PlaylistsHelper.MAX_CONCURRENT_IMPORT_CALLS
+import com.github.libretube.api.obj.Playlist
+import com.github.libretube.api.obj.Playlists
+import com.github.libretube.api.obj.StreamItem
+import com.github.libretube.api.obj.Subscription
+import com.github.libretube.db.DatabaseHolder.Database
+import com.github.libretube.db.obj.LocalPlaylist
+import com.github.libretube.db.obj.LocalSubscription
+import com.github.libretube.extensions.TAG
+import com.github.libretube.extensions.parallelMap
+import com.github.libretube.obj.PipedImportPlaylist
+import com.github.libretube.repo.LocalFeedRepository.Companion.CHANNEL_BATCH_DELAY
+import com.github.libretube.repo.LocalFeedRepository.Companion.CHANNEL_BATCH_SIZE
+import com.github.libretube.repo.LocalFeedRepository.Companion.CHANNEL_CHUNK_SIZE
+import com.github.libretube.ui.dialogs.ShareDialog.Companion.YOUTUBE_FRONTEND_URL
+import kotlinx.coroutines.delay
+import org.schabi.newpipe.extractor.channel.ChannelInfo
+import java.util.concurrent.atomic.AtomicInteger
+
+class LocalUserDataRepository : UserDataRepository {
+    override var requiresLogin: Boolean = false
+
+    override suspend fun getPlaylist(playlistId: String): Playlist {
+        val relation = Database.localPlaylistsDao().getAll()
+            .first { it.playlist.id.toString() == playlistId }
+
+        return Playlist(
+            name = relation.playlist.name,
+            description = relation.playlist.description,
+            thumbnailUrl = relation.playlist.thumbnailUrl,
+            videos = relation.videos.size,
+            relatedStreams = relation.videos.map { it.toStreamItem() }
+        )
+    }
+
+    override suspend fun getPlaylists(): List<Playlists> {
+        return Database.localPlaylistsDao().getAll()
+            .map {
+                Playlists(
+                    id = it.playlist.id.toString(),
+                    name = it.playlist.name,
+                    shortDescription = it.playlist.description,
+                    thumbnail = it.playlist.thumbnailUrl,
+                    videos = it.videos.size.toLong()
+                )
+            }
+    }
+
+    override suspend fun addToPlaylist(playlistId: String, vararg videos: StreamItem): Boolean {
+        val localPlaylist = Database.localPlaylistsDao().getAll()
+            .first { it.playlist.id.toString() == playlistId }
+
+        for (video in videos) {
+            val localPlaylistItem = video.toLocalPlaylistItem(playlistId)
+
+            val existingVideo = Database.localPlaylistsDao()
+                .getPlaylistVideo(playlistId, localPlaylistItem.videoId)
+            if (existingVideo != null) {
+                // update existing video metadata
+                localPlaylistItem.id = existingVideo.id
+                Database.localPlaylistsDao().updatePlaylistVideo(localPlaylistItem)
+                continue
+            }
+
+            // add the new video to the database
+            Database.localPlaylistsDao().addPlaylistVideo(localPlaylistItem)
+
+            val playlist = localPlaylist.playlist
+            if (playlist.thumbnailUrl.isEmpty()) {
+                // set the new playlist thumbnail URL
+                localPlaylistItem.thumbnailUrl?.let {
+                    playlist.thumbnailUrl = it
+                    Database.localPlaylistsDao().updatePlaylist(playlist)
+                }
+            }
+        }
+
+        return true
+    }
+
+    override suspend fun renamePlaylist(playlistId: String, newName: String): Boolean {
+        val playlist = Database.localPlaylistsDao().getAll()
+            .first { it.playlist.id.toString() == playlistId }.playlist
+        playlist.name = newName
+        Database.localPlaylistsDao().updatePlaylist(playlist)
+
+        return true
+    }
+
+    override suspend fun changePlaylistDescription(playlistId: String, newDescription: String): Boolean {
+        val playlist = Database.localPlaylistsDao().getAll()
+            .first { it.playlist.id.toString() == playlistId }.playlist
+        playlist.description = newDescription
+        Database.localPlaylistsDao().updatePlaylist(playlist)
+
+        return true
+    }
+
+    override suspend fun clonePlaylist(playlistId: String): String {
+        val playlist = MediaServiceRepository.instance.getPlaylist(playlistId)
+        val newPlaylist = createPlaylist(playlist.name ?: "Unknown name")
+
+        PlaylistsHelper.addToPlaylist(newPlaylist, *playlist.relatedStreams.toTypedArray())
+
+        var nextPage = playlist.nextpage
+        while (nextPage != null) {
+            nextPage = runCatching {
+                MediaServiceRepository.instance.getPlaylistNextPage(playlistId, nextPage).apply {
+                    PlaylistsHelper.addToPlaylist(newPlaylist, *relatedStreams.toTypedArray())
+                }.nextpage
+            }.getOrNull()
+        }
+
+        return playlistId
+    }
+
+    override suspend fun removeFromPlaylist(playlistId: String, index: Int): Boolean {
+        val transaction = Database.localPlaylistsDao().getAll()
+            .first { it.playlist.id.toString() == playlistId }
+        Database.localPlaylistsDao().removePlaylistVideo(
+            transaction.videos[index]
+        )
+        // set a new playlist thumbnail if the first video got removed
+        if (index == 0) {
+            transaction.playlist.thumbnailUrl =
+                transaction.videos.getOrNull(1)?.thumbnailUrl.orEmpty()
+        }
+        Database.localPlaylistsDao().updatePlaylist(transaction.playlist)
+
+        return true
+    }
+
+    override suspend fun importPlaylists(playlists: List<PipedImportPlaylist>) {
+        for (playlist in playlists) {
+            val playlistId = createPlaylist(playlist.name!!)
+
+            // if not logged in, all video information needs to become fetched manually
+            // Only do so with `MAX_CONCURRENT_IMPORT_CALLS` videos at once to prevent performance issues
+            for (videoIdList in playlist.videos.chunked(MAX_CONCURRENT_IMPORT_CALLS)) {
+                val streams = videoIdList.parallelMap {
+                    runCatching { MediaServiceRepository.instance.getStreams(it) }
+                        .getOrNull()
+                        ?.toStreamItem(it)
+                }.filterNotNull()
+
+                PlaylistsHelper.addToPlaylist(playlistId, *streams.toTypedArray())
+            }
+        }
+    }
+
+    override suspend fun createPlaylist(playlistName: String): String {
+        val playlist = LocalPlaylist(name = playlistName, thumbnailUrl = "")
+        return Database.localPlaylistsDao().createPlaylist(playlist).toString()
+    }
+
+    override suspend fun deletePlaylist(playlistId: String): Boolean {
+        Database.localPlaylistsDao().deletePlaylistById(playlistId)
+        Database.localPlaylistsDao().deletePlaylistItemsByPlaylistId(playlistId)
+
+        return true
+    }
+
+    override suspend fun subscribe(
+        channelId: String, name: String, uploaderAvatar: String?, verified: Boolean
+    ) {
+        val localSubscription = LocalSubscription(
+            channelId = channelId,
+            name = name,
+            avatar = uploaderAvatar,
+            verified = verified
+        )
+
+        Database.localSubscriptionDao().insert(localSubscription)
+    }
+
+    override suspend fun unsubscribe(channelId: String) {
+        Database.localSubscriptionDao().deleteById(channelId)
+    }
+
+    override suspend fun isSubscribed(channelId: String): Boolean {
+        return Database.localSubscriptionDao().includes(channelId)
+    }
+
+    override suspend fun importSubscriptions(newChannels: List<String>) {
+        val subscribedChannels = getSubscriptionChannelIds()
+
+        val newFiltered = newChannels.filter { !subscribedChannels.contains(it) }
+
+        val failedChannels = mutableListOf<String>()
+
+        val channelExtractionCount = AtomicInteger()
+        for (chunk in newFiltered.chunked(CHANNEL_CHUNK_SIZE)) {
+            // avoid being rate-limited by adding random delays between requests
+            val count = channelExtractionCount.get()
+            if (count >= CHANNEL_BATCH_SIZE) {
+                // add a delay after each BATCH_SIZE amount of fully-fetched channels
+                delay(CHANNEL_BATCH_DELAY.random())
+                channelExtractionCount.set(0)
+            }
+
+            chunk.parallelMap { channelId ->
+                try {
+                    val channelUrl = "$YOUTUBE_FRONTEND_URL/channel/${channelId}"
+                    val channelInfo = ChannelInfo.getInfo(channelUrl)
+
+                    val avatarUrl = channelInfo.avatars.maxByOrNull { it.height }?.url
+                    subscribe(channelId, channelInfo.name, avatarUrl, channelInfo.isVerified)
+                } catch (e: Exception) {
+                    Log.e(TAG(), e.toString())
+                    failedChannels.add(channelId)
+                }
+            }
+        }
+
+        if (!failedChannels.isEmpty()) {
+            throw Exception("Failed to import ${failedChannels.joinToString(", ")}")
+        }
+    }
+
+    override suspend fun getSubscriptions(): List<Subscription> {
+        // load all channels that have not been fetched yet
+        val unfinished = Database.localSubscriptionDao().getChannelsWithoutMetaInfo()
+        runCatching {
+            importSubscriptions(unfinished.map { it.channelId })
+        }
+
+        return Database.localSubscriptionDao().getAll().map {
+            Subscription(
+                url = it.channelId,
+                name = it.name.orEmpty(),
+                avatar = it.avatar,
+                verified = it.verified
+            )
+        }
+    }
+
+    override suspend fun getSubscriptionChannelIds(): List<String> {
+        return Database.localSubscriptionDao().getAll().map { it.channelId }
+    }
+
+    override suspend fun submitSubscriptionChannelInfosChanged(subscriptions: List<Subscription>) {
+        Database.localSubscriptionDao().updateAll(subscriptions.map {
+            LocalSubscription(
+                it.url,
+                it.name,
+                it.avatar,
+                it.verified
+            )
+        })
+    }
+}
