+++ app/src/main/java/com/github/libretube/repo/UserDataRepository.kt
@@ -3,9 +3,24 @@ package com.github.libretube.repo
 import com.github.libretube.api.obj.Playlist
 import com.github.libretube.api.obj.Playlists
 import com.github.libretube.api.obj.StreamItem
+import com.github.libretube.api.obj.Subscription
 import com.github.libretube.obj.PipedImportPlaylist
 
-interface PlaylistRepository {
+interface UserDataRepository {
+    var requiresLogin: Boolean
+
+    suspend fun login(username: String, password: String): String = ""
+    suspend fun register(username: String, password: String): String = ""
+    suspend fun deleteAccount(password: String) = Unit
+
+    suspend fun subscribe(channelId: String, name: String, uploaderAvatar: String?, verified: Boolean)
+    suspend fun unsubscribe(channelId: String)
+    suspend fun isSubscribed(channelId: String): Boolean?
+    suspend fun importSubscriptions(newChannels: List<String>)
+    suspend fun getSubscriptions(): List<Subscription>
+    suspend fun getSubscriptionChannelIds(): List<String>
+    suspend fun submitSubscriptionChannelInfosChanged(subscriptions: List<Subscription>) {}
+
     suspend fun getPlaylist(playlistId: String): Playlist
     suspend fun getPlaylists(): List<Playlists>
     suspend fun addToPlaylist(playlistId: String, vararg videos: StreamItem): Boolean
